Unifying the elasticsearch hooks into a single file
