+++ airflow/providers/elasticsearch/provider.yaml
@@ -55,9 +55,6 @@ hooks:
   - integration-name: Elasticsearch
     python-modules:
       - airflow.providers.elasticsearch.hooks.elasticsearch
-  - integration-name: Elasticsearch Hook (Python Client)
-    python-modules:
-      - airflow.providers.elasticsearch.hooks.elasticsearch_python
 
 connection-types:
   - hook-class-name: airflow.providers.elasticsearch.hooks.elasticsearch.ElasticsearchHook
