+++ tests/providers/elasticsearch/hooks/test_elasticsearch.py
@@ -20,8 +20,10 @@
 import unittest
 from unittest import mock
 
+from elasticsearch import Elasticsearch
+
 from airflow.models import Connection
-from airflow.providers.elasticsearch.hooks.elasticsearch import ElasticsearchHook
+from airflow.providers.elasticsearch.hooks.elasticsearch import ElasticsearchHook, ElasticsearchPythonHook
 
 
 class TestElasticsearchHookConn(unittest.TestCase):
@@ -95,3 +97,33 @@ class TestElasticsearchHook(unittest.TestCase):
         assert result_sets[1][0] == df.values.tolist()[1][0]
 
         self.cur.execute.assert_called_once_with(statement)
+
+
+class MockElasticsearch:
+    def __init__(self, data: dict):
+        self.data = data
+
+    def search(self, **kwargs):
+        return self.data
+
+
+class TestElasticsearchPythonHook:
+    def setup(self):
+        self.elasticsearch_hook = ElasticsearchPythonHook(hosts=["http://localhost:9200"])
+
+    def test_client(self):
+        es_connection = self.elasticsearch_hook.get_conn()
+        assert isinstance(es_connection, Elasticsearch)
+
+    @mock.patch("airflow.providers.elasticsearch.hooks.elasticsearch.ElasticsearchPythonHook.get_conn")
+    def test_search(self, elastic_mock):
+        es_data = {"hits": "test_hit"}
+        es_client = MockElasticsearch(es_data)
+        elastic_mock.return_value = es_client
+        print(MockElasticsearch.search)
+
+        query = {"test_query": "test_filter"}
+
+        result = self.elasticsearch_hook.search(index="test_index", query=query)
+
+        assert result == es_data['hits']
